# docs/openapi3.yaml
openapi: 3.0.0
...
info:
  contact:
    email: shpakovskiipa@gmail.com
    name: API Support
  description: This is a Eschool API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Eschool API
  version: "1.0"
paths:
...
  /courses:
    get:
      description: get all courses
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                    http_v1_dto.CourseDTO"
                type: array
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      summary: GetAllCourses
      tags:
        - course
  "/courses/{courseID}/lessons/{lessonID}":
    get:
      description: get lesson by id
      parameters:
        - description: course id
          in: path
          name: courseID
          required: true
          schema:
            type: string
        - description: lesson id
          in: path
          name: lessonID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                    http_v1_dto.LessonDTO"
                type: array
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorNotFound"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      security:
        - ApiKeyAuth: []
      summary: GetLessonByID
      tags:
        - course
    put:
      description: update course lesson
      parameters:
        - description: course id
          in: path
          name: courseID
          required: true
          schema:
            type: string
        - description: lesson id
          in: path
          name: lessonID
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                http_v1_dto.UpdateLessonDTO"
        description: updated lesson info
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                  http_v1_dto.LessonDTO"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorUnauthor\
                  ized"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorNotFound"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      security:
        - ApiKeyAuth: []
      summary: UpdateCourseLesson
      tags:
        - course
  "/courses/{courseID}/teachers/{teacherID}":
    put:
      description: add course teacher
      parameters:
        - description: course id
          in: path
          name: courseID
          required: true
          schema:
            type: string
        - description: course id
          in: path
          name: teacherID
          required: true
          schema:
            type: string
      responses:
        "201":
          description: message
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorUnauthor\
                  ized"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorNotFound"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      security:
        - ApiKeyAuth: []
      summary: AddCourseTeacher
      tags:
        - course
  "/courses/{id}":
    get:
      description: get course by id
      parameters:
        - description: course id
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                    http_v1_dto.CourseDTO"
                type: array
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorNotFound"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      summary: GetCourseByID
      tags:
        - course
  "/courses/{id}/lessons":
    get:
      description: get course lessons
      parameters:
        - description: course id
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                    http_v1_dto.LessonDTO"
                type: array
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorUnauthor\
                  ized"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorNotFound"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      security:
        - ApiKeyAuth: []
      summary: GetCourseLessons
      tags:
        - course
    post:
      description: create course lesson
      parameters:
        - description: course id
          in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                http_v1_dto.CreateLessonDTO"
        description: created lesson info
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                  http_v1_dto.LessonDTO"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorUnauthor\
                  ized"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorNotFound"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      security:
        - ApiKeyAuth: []
      summary: CreateCourseLesson
      tags:
        - course
  "/courses/{id}/teachers":
    get:
      description: get course teachers
      parameters:
        - description: course id
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                    http_v1_dto.UserDTO"
                type: array
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorNotFound"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      summary: GetCourseTeachers
      tags:
        - course
...
  "/schools/{courseID}/courses/{lessonID}":
    delete:
      description: delete course lesson
      parameters:
        - description: course id
          in: path
          name: courseID
          required: true
          schema:
            type: string
        - description: lesson id
          in: path
          name: lessonID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: message
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorUnauthor\
                  ized"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorNotFound"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      security:
        - ApiKeyAuth: []
      summary: DeleteCourseLesson
      tags:
        - course
...
  /users/courses:
    get:
      description: find user courses
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                    http_v1_dto.CourseDTO"
                type: array
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorUnauthor\
                  ized"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      security:
        - ApiKeyAuth: []
      summary: FindUserCourses
      tags:
        - user
...
servers:
  - url: //localhost:8080/api/v1
components:
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: Authorization
      type: apiKey
  schemas:
...
    internal_adapter_delivery_http_v1.RestErrorNotFound:
      properties:
        error:
          example: not found
          type: string
        status:
          example: 404
          type: integer
        timestamp:
          example: 2020-11-10T23:00:00+00:00
          type: string
      type: object
...